exc hw 002

+++ src/exceptions/homework002/Task002.java
@@ -13,7 +13,7 @@ try {
 
 public class Task002 {
     public static void main(String[] args) {
-        float[] intArray = new float[9];
+        int[] intArray = new int[9];
         try {
             int d = 0;
             double catchedRes1 = intArray[8] / d;
